Implement anomalous structure factor output to MTZ files in sfcalc
Add support for writing anomalous structure factors (FCanom and PHICanom columns) to MTZ files alongside regular structure factors. Anomalous data is calculated using f" coefficients via FFT and written to configurable column names.

- Add --fcanom and --phicanom options for custom column labels
- Calculate anomalous structure factors using ZeroCoef template with f"
- Output both regular (FC/PHIC) and anomalous data in single MTZ file
- Add validation for wavelength, unit cell, and spacegroup compatibility

+++ include/gemmi/neutron92.hpp
@@ -26,7 +26,7 @@ struct ZeroCoef {
   static Real data[121];
 
   static Real& get_(El el) { return data[static_cast<int>(el)]; }
-  static bool has(El el) { return static_cast<int>(el) < sizeof(data) / sizeof(Real); }
+  static bool has(El el) { return static_cast<size_t>(el) < sizeof(data) / sizeof(Real); }
   static Coef get(El el, signed char /*charge*/=0, int /*serial*/=0) { return Coef{{get_(el)}}; }
 };
